+++ CHANGES.md
@@ -2,9 +2,18 @@
 
 ## O.4.2
 
+### Improvements
+
+-   new `get_gdal_data_path()` utility funtion to check the path of the data
+    directory detected by GDAL (#160)
+
 ### Bug fixes
 
 -   register GDAL drivers during initial import of pyogrio (#145)
+-   support writing "not a time" (NaT) values in a datetime column (#146)
+-   fixes an error when reading GPKG with bbox filter (#150)
+-   properly raises error when invalid where clause is used on a GPKG (#150)
+-   avoid duplicate count of available features (#151)
 
 ## 0.4.1